Improve documentation
This change aims to improve on the grammar in the comments.
I also adjusted the Personal Goals link in the README
to match the order in the detailed section.

These are not very critical changes. As a newbie to Rust,the goal
for me here is to get familiar with the project as I find
it particularly interesting :)